change user entity id to long change data base column userId int(11) to bigint(20)

+++ searchmytraining/src/main/java/com/searchmytraining/dao/IEmploymentDAO.java
@@ -5,6 +5,6 @@ import com.searchmytraining.entity.EmploymentEntity;
 public interface IEmploymentDAO {
 	
 	public void updateEmpldet(EmploymentEntity emplEntity);
-	public EmploymentEntity findEmplDet(Integer userid);
+	public EmploymentEntity findEmplDet(Long userid);
 
 }
