+++ searchmytraining/src/main/java/com/searchmytraining/service/ITrainingProviderService.java
@@ -4,7 +4,7 @@ import com.searchmytraining.dto.TrainerDTO;
 import com.searchmytraining.entity.TrainerEntity;
 
 public interface ITrainingProviderService {
-		public Integer registerTrainer(TrainerDTO trainerdto);
+		public Long registerTrainer(TrainerDTO trainerdto);
 		public TrainerEntity getTrainer(Long id);
 		public TrainerEntity getTrainerByUserid(Long userid);
 }
